% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBetaBin.R
\name{calcBetaBin}
\alias{calcBetaBin}
\title{Calculate reliability using a Beta-Binomial model}
\usage{
calcBetaBin(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity}

\item{y}{variable to use as the outcome}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
Estimated parameters from the Beta-Binomial model, estimates of between and within-entity variance, and estimates of entity-specific reliability
}
\description{
This function estimates reliability using a Beta-Binomial model.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
