% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHLGMRel.R
\name{calcHLGMRel}
\alias{calcHLGMRel}
\title{Calculate reliability using a hierarchical logistic regression model}
\usage{
calcHLGMRel(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  show.all = FALSE,
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\description{
This function estimates reliability using a hierarchical logistic regression model.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
