% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcReliability.R
\name{calcReliability}
\alias{calcReliability}
\title{Calculate reliability of quality measure performance}
\usage{
calcReliability(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  show.all = FALSE,
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
A list with the following components:
\item{perf.res}{list of measure performance results}
\item{rel.results}{table of reliability estimates by method}
}
\description{
This function calculates several estimates of quality measure performance.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
