% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colonoscopy}
\alias{colonoscopy}
\title{Colonoscopy follow-up data}
\format{
A data frame with five variables:
\code{entity}, \code{p},
\code{n}, \code{x}.
}
\usage{
colonoscopy
}
\description{
These data are from the Centers for Medicare and Medicaid Services (CMS). They contain the percentage of patients receiving appropriate recommendation for follow-up screening colonoscopy
in calendar year 2023.
}
\details{
Data can be accessed from this link: \link{https://data.cms.gov/provider-data/dataset/yv7e-xc69}

\code{entity} indicates the accountable health care entity;
\code{p} indicates the follow-up rate;
\code{n} indicates the denominator count for the measure;
\code{x} indicates the numerator count for the measure.
}
\keyword{datasets}
