% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlPerf.R
\name{controlPerf}
\alias{controlPerf}
\title{Parameters for performance calculations}
\usage{
controlPerf(min.n = 2, alpha = 0.05, n.boots = 1000, n.cores = 2)
}
\arguments{
\item{min.n}{minimum number of observations per entity}

\item{alpha}{statistical significance level to use for confidence intervals}

\item{n.boots}{number of bootstraps to use for confidence interval estimation for P/E ratios}

\item{n.cores}{number of cores to use for parallel processing}
}
\description{
This function stores parameters for performance calculations
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
