% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estOEPE.R
\name{estOEPE}
\alias{estOEPE}
\title{Calculate measure performance by accountable entity}
\usage{
estOEPE(df, model, entities, entity = "entity", y = "y")
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entities}{list of accountable entities}

\item{entity}{variable to use as the accountable entity; default = "entity"}

\item{y}{variable to use as the outcome; default = "y"}
}
\value{
Estimated risk-standardized measure performance by accountable entity
}
\description{
This function calculates measure performance by accountable entity.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
