% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df_1}
\alias{example_df_1}
\title{Example data set used in tutorial}
\format{
A data frame with six variables:
\code{entity}, \code{z}, \code{x1}, \code{lp}, \code{p}, \code{y}.
}
\usage{
example_df_1
}
\description{
This is an example simulated data frame generated by \code{simulateData()} with
100 accountable entities, an average of 50 observations per entity, a marginal outcome rate
of 0.2, a median reliability of 0.7, and a regression coefficient of 0 for the covariate.
}
\keyword{datasets}
