% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Calculate model performance}
\usage{
model_performance(
  df,
  model,
  entity,
  y,
  predictor.clean,
  ctrPerf,
  output.dir,
  filename.add = NULL
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity; default = "entity"}

\item{y}{variable to use as the outcome; default = "y"}

\item{entities}{list of accountable entities}
}
\value{
Estimated risk-standardized measure performance by accountable entity
}
\description{
This function calculates risk model performance.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
