% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstimates.R
\name{plotEstimates}
\alias{plotEstimates}
\title{Plot estimates from regression model used for risk-adjustment}
\usage{
plotEstimates(model.performance = model.performance)
}
\description{
This function creates a plot of model results
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
