% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMisclassification.R
\name{plotMisclassification}
\alias{plotMisclassification}
\title{Plots misclassification visualization}
\usage{
plotMisclassification(entities, orig.est, boot.est, num.groups, out.file)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity; default = "entity"}

\item{y}{variable to use as the outcome; default = "y"}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
Estimated measure performance by accountable entity
}
\description{
This function plots a visualization of misclassification probability
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
