% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerformance.R
\name{plotPerformance}
\alias{plotPerformance}
\title{Plot measure performance across accountable entities}
\usage{
plotPerformance(df, plot.y = "p")
}
\arguments{
\item{df}{perf.results dataframe from calcPerformance() output}

\item{plot.y}{'p' plots the unadjusted performance, 'oe' plots the observed-to-expected ratio based on the risk-adjustment model}
}
\value{
A ggplot figure
}
\description{
This function creates a plot of measure performance across accountable entities, using the perf.results dataframe from calcPerformance() output.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
