% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReliability.R
\name{plotReliability}
\alias{plotReliability}
\title{Plot distributions of reliability estimates across entities}
\usage{
plotReliability(rel.out = rel.out)
}
\arguments{
\item{rel.out}{results from calcReliability()}
}
\description{
This function creates boxplots of reliability estimates across entities and different methods
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
