% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNRReliability.R
\name{plotSNRReliability}
\alias{plotSNRReliability}
\title{Plot distributions of reliability estimates across entities}
\usage{
plotSNRReliability(rel.out = rel.out)
}
\description{
This function creates boxplots of reliability estimates across entities and different methods
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
