% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSSR.R
\name{plotSSR}
\alias{plotSSR}
\title{Example plot of values used to calcualte split-sample reliability.}
\usage{
plotSSR(df, model = NULL, entity = "entity", y = "y", ctrPerf = controlPerf())
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\description{
This function creates a plot of the measure performance for each entity split across two exclusive random samples of observations within each entity.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
