% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_analysis.R
\name{profiling_analysis}
\alias{profiling_analysis}
\title{Calculate measure performance}
\usage{
profiling_analysis(
  df,
  model = NULL,
  entity = "entity",
  y = "y",
  data.type = "binary",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{observation-level data}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{data.type}{acceptable values are \code{binary} for 0/1 data and \code{continuous} for continuous data (default: \code{binary})}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
Estimated measure performance by accountable entity
}
\description{
This function calculates measure performance by accountable entity.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
