% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psychreadmission}
\alias{psychreadmission}
\title{Psychiatric 30-day, unplanned readmission data}
\format{
A data frame with six variables:
\code{entity}, \code{category},
\code{n}, \code{rate}, \code{rate.lwr}, \code{rate.upr}.
}
\usage{
psychreadmission
}
\description{
These data are from the Centers for Medicare and Medicaid Services (CMS) and show the percentage of patients who return to a hospital for an unplanned inpatient stay after discharge. Data are from 07/01/2021 to 06/30/2023
}
\keyword{datasets}
