% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_analysis.R
\name{reliability_analysis}
\alias{reliability_analysis}
\title{Calculate reliability}
\usage{
reliability_analysis(
  df,
  model,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
Estimated risk-standardized measure performance by accountable entity
}
\description{
This function calculates reliability using several methods.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
