% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tutorial_BB_agg_results}
\alias{tutorial_BB_agg_results}
\title{Beta-Binomial reliability results for example aggregated data set used in tutorial}
\format{
An object of class \code{list} of length 8.
}
\usage{
tutorial_BB_agg_results
}
\description{
This is the pre-computed reliability results used in the tutorial for the Beta-Binomial method
applied to an aggregated data set.
}
\keyword{datasets}
