% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tutorial_profiling_results}
\alias{tutorial_profiling_results}
\title{Risk-adjusted profiling results for example data set used in tutorial}
\format{
An object of class \code{data.frame} with 100 rows and 35 columns.
}
\usage{
tutorial_profiling_results
}
\description{
This is the pre-computed profiling results used in the tutorial for the risk-adjusted example.
}
\keyword{datasets}
