% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tutorial_reliability_results_2}
\alias{tutorial_reliability_results_2}
\title{Reliability results for example data set with risk adjustment used in tutorial}
\format{
An object of class \code{list} of length 7.
}
\usage{
tutorial_reliability_results_2
}
\description{
This is the pre-computed reliability results used in the tutorial for example 2.
}
\keyword{datasets}
