% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAOV.R
\name{calcAOV}
\alias{calcAOV}
\title{Calculate reliability using ANOVA method}
\usage{
calcAOV(df = NULL, model = NULL, y, provider, ctrPerf = controlPerf())
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{y}{variable to use as the outcome}

\item{provider}{variable to use as the accountable entity}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
A list with the following components:
\item{var.b.aov}{between-entity variance}
\item{var.w.aov}{within-entity variance}
\item{est.aov}{reliability estimate}

The plot function can be used to plot the provider-level reliability estimates.
}
\description{
This function estimates reliability using the ANOVA method.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
