% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAOV.R
\name{calcAOV}
\alias{calcAOV}
\title{Calculate reliability using one-way ANOVA method}
\usage{
calcAOV(
  df,
  entity = "entity",
  y = "y",
  df.aggregate = FALSE,
  n = "n",
  mean = "mean",
  std.dev = "sd",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{observation-level data}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{df.aggregate}{set this to \code{TRUE} if the data have already been aggregated to include only summary data (sample size, means, and standard deviations) for each entity; default is \code{FALSE}.}

\item{n}{if using aggregated data, data column containing the sample sizes for each entity; default is \code{n}.}

\item{mean}{if using aggregated data, data column containing the sample means for each entity; default is \code{mean}.}

\item{std.dev}{if using aggregated data, data column containing the sample standard deviations for each entity entity; default is \code{sd}.}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\description{
This function estimates reliability using the one-way ANOVA method.
}
\references{
He K, Kalbfleisch JD, Yang Y, Fei Z. Inter‐unit reliability for nonlinear models. Statistics in Medicine. 2019 Feb 28;38(5):844-54.

Nieser KJ, Harris AH. Comparing methods for assessing the reliability of health care quality measures. Statistics in Medicine. 2024 Oct 15;43(23):4575-94.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
