% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAOV.R
\name{calcAOV}
\alias{calcAOV}
\title{Calculate reliability using ANOVA method}
\usage{
calcAOV(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
A list with the following components:
\item{var.b.aov}{between-entity variance}
\item{var.w.aov}{within-entity variance}
\item{est.aov}{reliability estimate}
}
\description{
This function estimates reliability using the ANOVA method.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
