% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAOV.R
\name{calcAOV}
\alias{calcAOV}
\title{Calculate reliability using one-way ANOVA method}
\usage{
calcAOV(df, entity = "entity", y = "y", ctrPerf = controlPerf())
}
\arguments{
\item{df}{observation-level data}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
A list with the following components:
\item{var.b.aov}{between-entity variance}
\item{var.w.aov}{within-entity variance}
\item{est.aov}{reliability estimate}
}
\description{
This function estimates reliability using the one-way ANOVA method.
}
\references{
He K, Kalbfleisch JD, Yang Y, Fei Z. Inter‐unit reliability for nonlinear models. Statistics in Medicine. 2019 Feb 28;38(5):844-54.

Nieser KJ, Harris AH. Comparing methods for assessing the reliability of health care quality measures. Statistics in Medicine. 2024 Oct 15;43(23):4575-94.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
