% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBetaBin.R
\name{calcBetaBin}
\alias{calcBetaBin}
\title{Calculate reliability using a Beta-Binomial model}
\usage{
calcBetaBin(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  df.aggregate = FALSE,
  n = "n",
  x = "x",
  show.all = FALSE,
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{dataframe (assumed to be observation-level unless \code{df.aggregate} is changed below); if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model (NOTE: currently, Beta-Binomial reliability estimates do not take risk-adjustment into account.)}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{df.aggregate}{set this to \code{TRUE} if the data have already been aggregated to include only summary data (sample sizes and numerators) for each entity; default is \code{FALSE}.}

\item{n}{if using aggregated data, data column containing the sample size by entity}

\item{x}{if using aggregated data, data column containing the number of observations that met measure criteria by entity}

\item{show.all}{logical parameter indicating whether all variations of reliability estimates should be calculated; default is \code{FALSE}.}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
A list containing estimates of Beta-Binomial model parameters (\code{alpha} and \code{beta}); between-entity variance (\code{var.between}), within-entity variance (\code{var.within}), and reliability (\code{est.BB}).
}
\description{
This function estimates reliability using a Beta-Binomial model. NOTE: currently, Beta-Binomial reliability estimates do not take risk-adjustment into account.
}
\examples{
# Simulate data
df <- simulateData(n.entity = 50, n.obs = 100, mu = .2, r = .7)

# Calculate reliability
out <- calcBetaBin(df = df, entity = 'entity', y = 'y')
summary(out$est.BB)

## Reliability can also be calculated with data aggregated by entity
df.agg <- data.frame(
          entity = aggregate(y ~ entity, data = df, length)$entity
          n = aggregate(y ~ entity, data = df, length)$y,
          x = aggregate(y ~ entity, data = df, sum)$y
          )

out2 <- calcBetaBin(df = df.agg, df.aggregate = T, n = 'n', x = 'x')
summary(out2$est.BB)

}
\references{
Adams JL. The Reliability of Provider Profiling: A Tutorial. 2009.

Nieser KJ, Harris AH. Comparing methods for assessing the reliability of health care quality measures. Statistics in Medicine. 2024 Oct 15;43(23):4575-94.

Zhou G, Lin Z. Improved beta-binomial estimation for reliability of healthcare quality measures. medRxiv. 2023 Jan 9:2023-01.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
