% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBetaBin.R
\name{calcBetaBin}
\alias{calcBetaBin}
\title{Calculate reliability using a Beta-Binomial model}
\usage{
calcBetaBin(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  df.aggregate = FALSE,
  n = "n",
  x = "x",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model (NOTE: currently, Beta-Binomial reliability estimates do not take risk-adjustment into account.)}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{df.aggregate}{set this to TRUE if the data have already been aggregated to include only summary data (n, x) for each entity; default is FALSE.}

\item{n}{if using aggregated data, data column containing the sample size by entity}

\item{x}{if using aggregated data, data column containing the number of observations that met measure criteria by entity}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
Estimated parameters from the Beta-Binomial model, estimates of between and within-entity variance, and estimates of entity-specific reliability
}
\description{
This function estimates reliability using a Beta-Binomial model.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
