% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHLGMRel.R
\name{calcHLGMRel}
\alias{calcHLGMRel}
\title{Calculate reliability using a multilevel logistic regression model}
\usage{
calcHLGMRel(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity}

\item{y}{variable to use as the outcome}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
Estimated parameters and reliability
\item{var.b.aov}{between-entity variance}
}
\description{
This function estimates reliability using a multilevel logistic regression model.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
