% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerformance.R
\name{calcPerformance}
\alias{calcPerformance}
\title{Calculates measure performance by accountable entity}
\usage{
calcPerformance(df = NULL, model = NULL, y, provider, ctrPerf = controlPerf())
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{y}{variable to use as the outcome}

\item{provider}{variable to use as the accountable entity}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
A list with the following components:
\item{var.b.aov}{between-entity variance}

The plot function can be used to plot the provider-level reliability estimates.
}
\description{
This function calculates measure performance by accountable entity.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
