% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerformance.R
\name{calcPerformance}
\alias{calcPerformance}
\title{Calculate measure performance by accountable entity}
\usage{
calcPerformance(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  data.type = "binary",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{data.type}{acceptable values are \code{binary} for 0/1 data (default: \code{binary})}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\description{
This function calculates measure performance by accountable entity.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
