% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcReliability.R
\name{calcReliability}
\alias{calcReliability}
\title{Calculate reliability of quality measure performance}
\usage{
calcReliability(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity; default = "entity"}

\item{y}{variable to use as the outcome; default = "y"}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
A list with the following components:
\item{perf.res}{list of measure performance results}
\item{rel.results}{table of reliability estimates by method}
}
\description{
This function calculates several estimates of quality measure performance.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
