% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResamplingIUR.R
\name{calcResamplingIUR}
\alias{calcResamplingIUR}
\title{Calculate reliability using resampling inter-unit reliability method}
\usage{
calcResamplingIUR(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity}

\item{y}{variable to use as the outcome}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
A list with the following components:
\item{var.b}{between-entity variance}
\item{var.w}{within-entity variance}
\item{var.total}{total variance}
\item{IUR}{estimate of IUR}
}
\description{
This function estimates reliability using the resampling inter-unit reliability method.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
