% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResamplingIUR.R
\name{calcResamplingIUR}
\alias{calcResamplingIUR}
\title{Calculate reliability using resampling inter-unit reliability method}
\usage{
calcResamplingIUR(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
A list with the following components:
\item{var.b}{between-entity variance}
\item{var.w}{within-entity variance}
\item{var.total}{total variance}
\item{IUR}{estimate of IUR}
}
\description{
This function estimates reliability using the resampling inter-unit reliability method.
}
\references{
He K, Kalbfleisch JD, Yang Y, Fei Z. Inter‐unit reliability for nonlinear models. Statistics in Medicine. 2019 Feb 28;38(5):844-54.

Nieser KJ, Harris AH. Comparing methods for assessing the reliability of health care quality measures. Statistics in Medicine. 2024 Oct 15;43(23):4575-94.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
