% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSR.R
\name{calcSSR}
\alias{calcSSR}
\title{Calculate reliability using split-sample method}
\usage{
calcSSR(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  data.type = "binary",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{data.type}{acceptable values are "binary" for 0/1 data and "continuous" for continuous data (default: 'binary')}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
SSR reliability estimates
}
\description{
This function estimates reliability using the split-sample method.
}
\references{
Nieser KJ, Harris AH. Split‐sample reliability estimation in health care quality measurement: Once is not enough. Health Services Research. 2024 Aug;59(4):e14310.

Nieser KJ, Harris AH. Comparing methods for assessing the reliability of health care quality measures. Statistics in Medicine. 2024 Oct 15;43(23):4575-94.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
