% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSR.R
\name{calcSSR}
\alias{calcSSR}
\title{Calculate reliability using split-sample method}
\usage{
calcSSR(
  df = NULL,
  model = NULL,
  y = "y",
  provider = "provider",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{y}{variable to use as the outcome}

\item{provider}{variable to use as the accountable entity}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
A list with the following components:
\item{var.b.aov}{between-entity variance}

The plot function can be used to plot the provider-level reliability estimates.
}
\description{
This function estimates reliability using the split-sample method.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
