% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSR.R
\name{calcSSR}
\alias{calcSSR}
\title{Calculate reliability using split-sample method}
\usage{
calcSSR(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
SSR reliability estimates
}
\description{
This function estimates reliability using the split-sample method.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
