% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlPerf.R
\name{controlPerf}
\alias{controlPerf}
\title{Parameters for performance calculations}
\usage{
controlPerf(min.n = 2, alpha = 0.05, n.boots = 200, n.cores = 4)
}
\description{
This function stores parameters for performance calculations
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
