% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlRel.R
\name{controlRel}
\alias{controlRel}
\title{Parameters for reliability calculations}
\usage{
controlRel(n.resamples = 100, SSRmethod = "permutation", fn = NA, d.steps = 10)
}
\description{
This function stores parameters for reliability calculations
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
