% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassification_analysis.R
\name{misclassification_analysis}
\alias{misclassification_analysis}
\title{Calculate misclassification probabilities}
\usage{
misclassification_analysis(
  df = NULL,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf(),
  ctrRel = controlRel()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}

\item{ctrRel}{parameters to control reliability estimation}
}
\value{
Estimated misclassification probabilities
}
\description{
This function runs the misclassification functions
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
