% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Calculate model performance}
\usage{
model_performance(
  df,
  model,
  entity = "entity",
  y = "y",
  data.type = "binary",
  predictor.clean = NA,
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{observation-level data; if null, will use the dataframe from the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{data.type}{acceptable values are "binary" for 0/1 data and "continuous" for continuous data (default: 'binary')}

\item{predictor.clean}{optional list of formatted names of predictors in the model}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
Estimated risk-standardized measure performance by accountable entity
}
\description{
This function calculates risk model performance.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
