% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalibration.R
\name{plotCalibration}
\alias{plotCalibration}
\title{Plot calibration curve for risk-adjustment model}
\usage{
plotCalibration(df = df, quantiles = 10)
}
\description{
This function creates a plot of the model calibration curve
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
