% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN.R
\name{plotN}
\alias{plotN}
\title{Plot sample size distribution across accountable entities}
\usage{
plotN(n)
}
\arguments{
\item{n}{vector of sample sizes}
}
\description{
This function creates a histogram of entity sample sizes.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
