% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredictedDistribution.R
\name{plotPredictedDistribution}
\alias{plotPredictedDistribution}
\title{Plot densities of predicted values by outcome group}
\usage{
plotPredictedDistribution(model.performance = model.performance)
}
\arguments{
\item{model.performance}{results from model_performance()}
}
\description{
This function creates a plot of the distributions of predicted values by outcome group
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
