% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReliability.R
\name{plotReliability}
\alias{plotReliability}
\title{Plot distributions of reliability estimates across entities}
\usage{
plotReliability(rel.out)
}
\arguments{
\item{rel.out}{results from \code{calcReliability()}}
}
\value{
A ggplot figure
}
\description{
This function creates boxplots of reliability estimates across entities and different methods
}
\examples{
### Simulate data with binary outcome
df <- simulateData(n.entity = 50, n.obs = 100, mu = .2, r = .7)

# Calculate reliability
out <- calcReliability(df = df, entity = 'entity', y = 'y')

# Plot estimates
plotReliability(out)


# Calculate reliability with expanded set of methods
out1 <- calcReliability(df = df, entity = 'entity', y = 'y', show.all = T)

# Plot estimates for expanded set of methods
plotReliability(out1)


### This function also works with continuous outcome data
# simulate data from multilevel normal distribution
df.c <- simulateData(n.entity = 50, n.obs = 100, mu = 25, r = .6, data.type = 'normal')

# calculate reliability
out.c <- calcReliability(df = df, entity = 'entity', y = 'y', data.type = 'continuous')
plotReliability(out.c)

}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
