% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_analysis.R
\name{profiling_analysis}
\alias{profiling_analysis}
\title{Calculate measure performance}
\usage{
profiling_analysis(
  df,
  model,
  entity,
  y,
  ctrPerf,
  output.dir,
  filename.add = NULL
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity; default = "entity"}

\item{y}{variable to use as the outcome; default = "y"}

\item{entities}{list of accountable entities}
}
\value{
Estimated risk-standardized measure performance by accountable entity
}
\description{
This function calculates measure performance by accountable entity.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
