% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_analysis.R
\name{profiling_analysis}
\alias{profiling_analysis}
\title{Calculate measure performance}
\usage{
profiling_analysis(
  df,
  model = NULL,
  entity = "entity",
  y = "y",
  ctrPerf = controlPerf()
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model (NOTE: currently, Beta-Binomial reliability estimates do not take risk-adjustment into account.)}

\item{entity}{data column containing the accountable entity identifier}

\item{y}{data column containing the outcome variable}

\item{ctrPerf}{parameters to control performance measure calculation}
}
\value{
Estimated risk-standardized measure performance by accountable entity
}
\description{
This function calculates measure performance by accountable entity.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
