% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_analysis.R
\name{reliability_analysis}
\alias{reliability_analysis}
\title{Calculate measure performance and reliability}
\usage{
reliability_analysis(
  df,
  model,
  entity,
  y,
  predictor.clean,
  ctrPerf,
  ctrRel,
  model.estimates.file,
  model.estimates.fig.file,
  prediction.fig.file,
  calibration.fig.file,
  rand.int.plot.file,
  corr.plot.file,
  comparison.rate.fig.file,
  rel.out.file,
  rel.dist.plot.file,
  SSR.plot.file
)
}
\arguments{
\item{df}{dataframe; if null, will use the dataframe in the model object}

\item{model}{model; if null, will use an unadjusted model}

\item{entity}{variable to use as the accountable entity; default = "entity"}

\item{y}{variable to use as the outcome; default = "y"}

\item{entities}{list of accountable entities}
}
\value{
Estimated risk-standardized measure performance by accountable entity
}
\description{
This function calculates measure performance by accountable entity.
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
