% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data}
\usage{
simulateData(
  n.providers,
  n.pts,
  type = "binary",
  p = NA,
  var.w = NA,
  lambda = NA
)
}
\arguments{
\item{n.providers}{the number of accountable entities}

\item{n.pts}{the number of observations for each accountable entity}

\item{type}{variable type of the observations (e.g., binary)}
}
\value{
A dataframe with the following columns:
\item{provider}{the accountable entity}
\item{y}{outcome}

The summary function can be used to...
}
\description{
This function simulates some data.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
