% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data}
\usage{
simulateData(n.entity, avg.n, icc, beta0, beta1, type = "binary")
}
\arguments{
\item{type}{variable type of the observations (e.g., binary)}

\item{n}{a vector of the number of observations for each accountable entity}
}
\value{
A dataframe with the following columns:
\item{entity}{the accountable entity}
\item{y}{outcome}
}
\description{
This function simulates some data.
}
\examples{
# TBD
}
\references{
None
}
\author{
Kenneth Nieser (nieser@stanford.edu)
}
